+++ bin/actions/0220022-45.sh
@@ -1,0 +1,18 @@
+#!/bin/bash
+
+#: ◣ + X in terminal - Type "exit" and hit ENTER
+
+focusApplication=\
+$(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
+
+path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 || exit ; pwd -P )"
+"$path_self"/utils/remove_unwanted_output.sh "$focusApplication"
+
+if [[ "$focusApplication" =~ "gnome-terminal-" ]]; then
+  xdotool sleep 0.1
+  xdotool type "exit"
+  xdotool key Return
+fi
+
+unset path_self
+unset focusApplication