+++ bin/actions/utils/remove_unwanted_output.sh
@@ -4,9 +4,13 @@
 # in applications showing it undesiredly
 # @param $1 identifier of focused application
 
-if [[ "$1" =~ soffice.bin ]] \
-  || [[ "$1" =~ "java" ]] \
+if [[ "$1" =~ "geany" ]] \
   || [[ "$1" =~ "gnome-terminal" ]] \
-  || [[ "$1" =~ "xpad" ]]; then
+  || [[ "$1" =~ "java" ]] \
+  || [[ "$1" =~ "mattermost-desk" ]] \
+  || [[ "$1" =~ soffice.bin ]] \
+  || [[ "$1" =~ "xpad" ]] \
+  || [[ "$1" =~ "yana" ]] \
+  ; then
     xdotool key 0xff08  # Backspace
 fi